// src/app/_models/rolename.enum.ts
export enum RoleName {
    Admin = 'Admin',
    Customer = 'Customer',
    Supplier = 'Supplier',

    SalesAdmin = 'SalesAdmin',
    SalesManager = 'SalesManager',
    SalesEmployee = 'SalesEmployee',

    CreditAdmin = 'CreditAdmin',
    CreditManager = 'CreditManager',
    CreditEmployee = 'CreditEmployee',

    LogisticsAdmin = 'LogisticsAdmin',
    LogisticsManager = 'LogisticsManager',
    LogisticsEmployee = 'LogisticsEmployee',

    MaintenanceAdmin = 'MaintenanceAdmin',
    MaintenanceManager = 'MaintenanceManager',
    MaintenanceEmployee = 'MaintenanceEmployee',

    ProductionAdmin = 'ProductionAdmin',
    ProductionManager = 'ProductionManager',
    ProductionEmployee = 'ProductionEmployee'
}